% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_baseflow.R
\name{average_baseflow}
\alias{average_baseflow}
\title{Baseflow Metric Function}
\usage{
average_baseflow(data)
}
\arguments{
\item{data}{a dataframe with hydrologic data}
}
\value{
The average baseflow
}
\description{
Calculates baseflow of dataframe - must be trimmed to water year
}
