% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drought_of_record.R
\name{drought_of_record}
\alias{drought_of_record}
\title{Drought of Record Metric Function}
\usage{
drought_of_record(data)
}
\arguments{
\item{data}{a dataframe containing hydrologic information}
}
\value{
The year when the Drought of Record occurs within dataframe
}
\description{
Calculates when the Drought of Record occurs in dataframe - must be trimmed to water year
}
