% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_exceedance.R
\name{flow_exceedance}
\alias{flow_exceedance}
\title{Flow Exceedance Probability Metric Function}
\usage{
flow_exceedance(data, prob)
}
\arguments{
\item{data}{a dataframe with hydrologic data}

\item{prob}{This will equal 0.01, 0.05, 0.5, 0.95, or 0.99}
}
\value{
The desired flow exceedance probability
}
\description{
Calculates a specific flow exceedance probability
}
