% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ALL_upstream.R
\name{fn_ALL_upstream}
\alias{fn_ALL_upstream}
\title{Upstream Segments Function}
\usage{
fn_ALL_upstream(riv.seg, AllSegList)
}
\arguments{
\item{riv.seg}{River segment of interest}

\item{AllSegList}{A list of all segments}
}
\value{
All upstream segments
}
\description{
Helps locate all upstream segments from the working segment
}
