% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_gage_and_seg_mapper.R
\name{fn_gage_and_seg_mapper}
\alias{fn_gage_and_seg_mapper}
\title{River Segment GIS Mapper Metric Function}
\usage{
fn_gage_and_seg_mapper(RivSeg, site, cbp6_link, token)
}
\arguments{
\item{RivSeg}{River segment ID}

\item{site}{USGS Gage ID}

\item{cbp6_link}{links function to correct computer directory}

\item{token}{allows access to deq2 website}
}
\value{
A GIS generated map of the river segment
}
\description{
Creates a GIS generated figure of the specific river segment
}
