% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_iha_7q10.R
\name{fn_iha_7q10}
\alias{fn_iha_7q10}
\title{7Q10 Metric Function}
\usage{
fn_iha_7q10(data)
}
\arguments{
\item{data}{a dataframe with hydrologic data}
}
\value{
The 7Q10 of the dataframe
}
\description{
Calculates 7Q10 of dataframe - must be trimmed to water year
}
