% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_iha_DOR.R
\name{fn_iha_DOR}
\alias{fn_iha_DOR}
\title{Drought of Record Flow Metric Function}
\usage{
fn_iha_DOR(data)
}
\arguments{
\item{data}{a dataframe with hydrologic data}
}
\value{
The drought of record flow
}
\description{
Calculates drought of record flow of dataframe - must be trimmed to water year
}
