% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_iha_DOR_Year.R
\name{fn_iha_DOR_Year}
\alias{fn_iha_DOR_Year}
\title{Drought of Record Year Metric Function}
\usage{
fn_iha_DOR_Year(data)
}
\arguments{
\item{data}{a dataframe with hydrologic data}
}
\value{
The drought of record year
}
\description{
Calculates drought of record year of dataframe - must be trimmed to water year
}
