% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_iha_mhf.R
\name{fn_iha_mhf}
\alias{fn_iha_mhf}
\title{Monthly High Flow Metric Function}
\usage{
fn_iha_mhf(data, month)
}
\arguments{
\item{data}{a dataframe with hydrologic data}

\item{month}{specified month of interest}
}
\value{
The monthly high flow
}
\description{
Calculates the monthly high flows of dataframe - must be trimmed to water year
}
