% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_iha_mlf.R
\name{fn_iha_mlf}
\alias{fn_iha_mlf}
\title{Monthly Low Flow Metric Function}
\usage{
fn_iha_mlf(data, month)
}
\arguments{
\item{data}{a dataframe with hydrologic data}

\item{month}{specified month of interest}
}
\value{
The monthly low flow
}
\description{
Calculates the monthly low flow of dataframe - must be trimmed to water year
}
