% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_upstream.R
\name{fn_upstream}
\alias{fn_upstream}
\title{Upstream Segment Function}
\usage{
fn_upstream(riv.seg, AllSegList)
}
\arguments{
\item{riv.seg}{River segment of interest}

\item{AllSegList}{List of all segments}
}
\value{
Upstream segment
}
\description{
Helps locate segments upstream of working segments
}
