% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeature.R
\name{getFeature}
\alias{getFeature}
\title{Get Feature Function}
\usage{
getFeature(inputs, token, base_url, feature)
}
\arguments{
\item{inputs}{the conveyance inputs}

\item{token}{token that allows you to access deq2}

\item{base_url}{computer directory to give R access to username and pw}

\item{feature}{feature of interest}
}
\value{
The features of the river segment
}
\description{
Collects information about the river segment of interest
}
