% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProperty.R
\name{getProperty}
\alias{getProperty}
\title{Get Property Function}
\usage{
getProperty(inputs, base_url)
}
\arguments{
\item{inputs}{the conveyance inputs}

\item{base_url}{computer directory to give R access to username and pw}
}
\value{
property of interest
}
\description{
pulls information about the river segment
}
