% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWatershedDF.R
\name{getWatershedDF}
\alias{getWatershedDF}
\title{Get Watershed Dataframe Function}
\usage{
getWatershedDF(geom)
}
\arguments{
\item{geom}{the geometry of interest}
}
\value{
A dataframe of the watershed
}
\description{
Retrieves the dataframe for the watershed
}
