% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage_import_data_cfs.R
\name{import_gage_data_cfs}
\alias{import_gage_data_cfs}
\title{Import USGS Gage Information Function}
\usage{
import_gage_data_cfs(siteNo, start.date, end.date)
}
\arguments{
\item{siteNo}{character USGS site number.  This is usually an 8 digit number.}

\item{start.date}{character starting date for data retrieval in the form YYYY-MM-DD.}

\item{end.date}{character ending date for data retrieval in the form YYYY-MM-DD.}
}
\value{
A dataframe containing the specfic gage's data for the specified time period
}
\description{
Imports data from the USGS, harvested from their stream gages
}
