% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_yearly_mean.R
\name{low_yearly_mean}
\alias{low_yearly_mean}
\title{Lowest Yearly Mean Metric Function}
\usage{
low_yearly_mean(data)
}
\arguments{
\item{data}{a dataframe with hydrologic data}
}
\value{
The lowest yearly mean flow
}
\description{
Calculates the lowest yearly mean flow of dataframe - must be trimmed to water year
}
