% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_calc_all.R
\name{metrics_calc_all}
\alias{metrics_calc_all}
\title{Baseflow Metric Function}
\usage{
metrics_calc_all(data)
}
\arguments{
\item{data}{a dataframe with hydrologic data}
}
\value{
A data frame with 65 different hydrologic metrics
}
\description{
Calculates 65 different hydrologic metrics from the given dataframe - must be trimmed to water year
}
