% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_compare.R
\name{metrics_compare}
\alias{metrics_compare}
\title{Percent Difference of Metrics Function}
\usage{
metrics_compare(metrics1, metrics2)
}
\arguments{
\item{metrics1}{a dataframe containing the 65 hydrologic metrics from dataframe 1}

\item{metrics2}{a dataframe containing the 65 hydrologic metrics from dataframe 2}
}
\value{
A dataframe containing the 65 metrics from both original dataframes, and the percent difference between each metric
}
\description{
Calculates percent difference between 65 different hydrologic metrics created from two different inputs
}
