% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_import_data_cfs.R
\name{model_import_data_cfs}
\alias{model_import_data_cfs}
\title{Import USGS Gage Information Function}
\usage{
model_import_data_cfs(riv.seg, mod.phase, mod.scenario, start.date,
  end.date)
}
\arguments{
\item{riv.seg}{Specific river segment of interest}

\item{mod.phase}{should be "p6/p6_gb604/tmp" (phase 6) or "p532c-sova" (phase 5)}

\item{mod.scenario}{should be "CFBASE30Y20180615" (phase 6) or "p532cal_062211"}

\item{start.date}{character starting date for data retrieval in the form YYYY-MM-DD.}

\item{end.date}{character ending date for data retrieval in the form YYYY-MM-DD.}
}
\value{
A dataframe containing the specfic river segments model data for the specified time period
}
\description{
Imports data from the USGS, harvested from their stream gages
}
