% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_mean.R
\name{monthly_mean}
\alias{monthly_mean}
\title{Monthly Mean Metric Function}
\usage{
monthly_mean(data, num.month)
}
\arguments{
\item{data}{dataframe of date and flow data taken from import data functions}

\item{num.month}{number for month of interest}
}
\value{
monthly mean metric
}
\description{
Measures monthly mean of dataframe - data must be trimmed to water year
}
