% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_min.R
\name{monthly_min}
\alias{monthly_min}
\title{Monthly Minimum Metric Function}
\usage{
monthly_min(data, num.month)
}
\arguments{
\item{data}{dataframe of date and flow data taken from import data functions}

\item{num.month}{number for month of interest}
}
\value{
monthly mean metric
}
\description{
Measures monthly minimum of dataframe - data must be trimmed to water year
}
