% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_mean_flow.R
\name{overall_mean_flow}
\alias{overall_mean_flow}
\title{Overall Mean Flow Metric Function}
\usage{
overall_mean_flow(data)
}
\arguments{
\item{data}{dataframe of date and flow data taken from import data functions}
}
\value{
overall mean flow metric
}
\description{
Measures the mean flow of the entire timeseries
}
