% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProperty.R
\name{postProperty}
\alias{postProperty}
\title{Post Property Function}
\usage{
postProperty(inputs, base_url, prop)
}
\arguments{
\item{inputs}{the conveyance inputs}

\item{base_url}{computer directory to give R access to username and pw}

\item{prop}{property code}
}
\value{
n/a
}
\description{
posts information about the river segment
}
