% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_token.R
\name{rest_token}
\alias{rest_token}
\title{Rest Token Function}
\usage{
rest_token(base_url, token, rest_uname = FALSE, rest_pw = FALSE)
}
\arguments{
\item{base_url}{computer directory to give R access to username and pw}

\item{token}{token that allows you to access deq2}

\item{rest_uname}{username to gain access to token}

\item{rest_pw}{password to gain access to token}
}
\value{
the rest token used to generate the gis image
}
\description{
Obtains token to access deq2
}
