% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sept_10_flow.R
\name{sept_10_flow}
\alias{sept_10_flow}
\title{September 10pct Metric Function}
\usage{
sept_10_flow(data)
}
\arguments{
\item{data}{dataframe of date and flow data taken from import data functions}
}
\value{
september 10pct metric
}
\description{
Measures the September 10pct of the entire timeseries

data must be trimmed to water year
}
