% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seven_q_ten.R
\name{seven_q_ten}
\alias{seven_q_ten}
\title{7Q10 Metric Function}
\usage{
seven_q_ten(data)
}
\arguments{
\item{data}{dataframe of date and flow data taken from import data functions}
}
\value{
7Q10 metric
}
\description{
Measures the 7Q10 of the entire timeseries - data must be trimmed to water year
}
