% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vahydro_import_all_metrics.R
\name{vahydro_import_all_metrics}
\alias{vahydro_import_all_metrics}
\title{Import All Metrics from VAHydro Function}
\usage{
vahydro_import_all_metrics(seg.or.gage, mod.scenario, token, site)
}
\arguments{
\item{seg.or.gage}{indicate segment name or gage number}

\item{mod.scenario}{input scenario code}

\item{token}{input token number for access}

\item{site}{input site name}
}
\value{
dataframe of all metrics
}
\description{
Imports all metrics from VAHydro for a given scenario
}
