% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vahydro_import_metric.R
\name{vahydro_import_metric}
\alias{vahydro_import_metric}
\title{Import Metric from VAHydro Function}
\usage{
vahydro_import_metric(met.varkey, met.propcode, seg.or.gage,
  mod.scenario = "p532cal_062211", token, site)
}
\arguments{
\item{met.varkey}{input variable key}

\item{met.propcode}{input prop code}

\item{seg.or.gage}{indicate segment name or gage number}

\item{mod.scenario}{input scenario code}

\item{token}{input token number for access}

\item{site}{input site name}
}
\value{
metric data
}
\description{
Imports metric from VAHydro for a given scenario
}
