% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vahydro_post_metric.R
\name{vahydro_post_metric}
\alias{vahydro_post_metric}
\title{Import Post Metric from VAHydro Function}
\usage{
vahydro_post_metric(met.varkey, met.propcode, met.name, met.value,
  seg.or.gage, mod.scenario = "p532cal_062211", token, site)
}
\arguments{
\item{met.varkey}{input variable key}

\item{met.propcode}{input prop code}

\item{met.name}{input name of metric}

\item{seg.or.gage}{indicate segment name or gage number}

\item{mod.scenario}{input scenario code}

\item{token}{input token number for access}

\item{site}{input site name}
}
\value{
post metric data
}
\description{
Imports post metric from VAHydro for a given scenario
}
