% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_year_trim.R
\name{water_year_trim}
\alias{water_year_trim}
\title{Trimming Data to Water Year Function}
\usage{
water_year_trim(data)
}
\arguments{
\item{data}{dataframe of date and flow data taken from import data functions}
}
\value{
dataframe trimmed to water year
}
\description{
Trims data to water year
}
